pyproject.toml
[tool.poetry]
name = "samwell"
version = "0.0.1"
description = "Useful utilities for biological data formats and analyses"
authors = ["Jeff Tratner <jeffrey.tratner@myriad.com>", "Nils Homer", "Tim Fennell"]

[tool.poetry.dependencies]
python = ">=3.6"
mypy-extensions = ">=0.4.3"
defopt = ">=5.1.0"
attrs = ">=19.3.0"
intervaltree = ">=3.0.2"
cython = ">=0.29.15"
pysam = ">=0.15.3"
pybedtools = ">=0.8.1"


[tool.poetry.dev-dependencies]
pytest = ">=5.3.5"
pytest-vcr = ">=1.0.2"
flake8 = ">=3.7.9"
mypy = ">=0.761"
pytest-cov = ">=2.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
